skip http/2 test in go1.5 and prior
